refactor taichi/common
